requirements.txt
bcrypt==3.2.2
uvicorn
fastapi
sqlalchemy
passlib
python-jose[cryptography]
python-multipart